Move version.rb under lib/langchain
