+++ lib/langchain/version.rb
